+++ testutils/src/main/java/com/google/android/exoplayer2/testutil/StubExoPlayer.java
@@ -300,6 +300,26 @@ public class StubExoPlayer extends BasePlayer implements ExoPlayer {
   }
 
   @Override
+  public void setFastForwardIncrement(long fastForwardIncrementMs) {
+    throw new UnsupportedOperationException();
+  }
+
+  @Override
+  public long getFastForwardIncrement() {
+    throw new UnsupportedOperationException();
+  }
+
+  @Override
+  public void setRewindIncrement(long rewindIncrementMs) {
+    throw new UnsupportedOperationException();
+  }
+
+  @Override
+  public long getRewindIncrement() {
+    throw new UnsupportedOperationException();
+  }
+
+  @Override
   public void setPlaybackParameters(PlaybackParameters playbackParameters) {
     throw new UnsupportedOperationException();
   }